Refactored the preprocess stage to clean up and make independent the various heuristics. Matter 2f2d
